Suggestion Overlay Improvements
Implement suggestion sorting - Suggestions will now be sorted by relevance. Previously, it was possible to type the exact characters for an emote, but have it not be a suggestion.

Implement query highlighting - The characters used in the search term will now be highlighted so that the user can understand why it's a recommendation. Also helps them parse which character they should type next.

Im