Optimize overlay
Noticed that with one or multiple large emoji displayed in the suggestions overlay, my client would slow down to a crawl. After some digging, I realized that the plugin was trying to load each image, normalize it, and display it on every frame.

Adding an image cache for the overlay caused the problem to disappear.